Refactor MVC pattern
It's been a while, university has not left me with much free time.
Anyway, a SharedLayout class with its own controller and model has
been created, it contains the shared elements both the searching and
sorting screens share. So sorting can be implemented.
