+++ screens/__init__.py
@@ -1,8 +1,8 @@
 from .screen import Screen
-from .introduction_screen.screen import Screen
-from .screen_template.screens_template import SharedLayout
+from .screen_template.screens_template import ScreenTemplate
 from .introduction_screen import IntroductionScreen
+from .shared_layout import SharedLayout
+from .shared_layout import SharedController
+from .shared_layout import SharedModel
+from .shared_layout import SharedDataModel
 from .search_screen import SearchScreen
-from .search_screen import SearchController
-from .search_screen import SearchModel
-from .search_screen import SearchDataModel
