+++ screens/screen_template/__init__.py
@@ -1,1 +1,1 @@
-from .screens_template import SharedLayout
+from .screens_template import ScreenTemplate
