+++ screens/search_screen/__init__.py
@@ -1,4 +1,3 @@
 from .searching_screen import SearchScreen
 from .search_controller import SearchController
 from .search_model import SearchModel
-from .data_model import SearchDataModel
