+++ screens/search_screen/searching_screen.py
@@ -7,298 +7,35 @@ if(__name__ == "__main__"):
     exit()
 
 import screens as sc
-from algorithms.handlers import callAlgorithm, getAlgorithms
 import tkinter as tk 
-from tkinter import ttk
-import random
-import threading
 
-class SearchScreen(sc.Screen, sc.SharedLayout):
+from .search_controller import SearchController
+from .search_model import SearchModel
+
+class SearchScreen(sc.Screen, sc.SharedLayout):    
     def initScreen(self) -> None:
-        # Creates basic layout of the screen
-        self.createTemplate()
-        # Override home button function
-        self.__overrideHomeButtonCommand()
-
-        # Dictionary pairing numbers to speed 
-        # This allows the slider to show text rather than just numbers
-        self.__numbersToText = {
-            0: "Target: Random", 
-            1: "Target: In array",
-            2: "Target: Not in array"
-        } 
-
-        # Dictionary pairing an integer to a function
-        self.__generateTargetFunctions = {
-            0 : self.targetRandom, 
-            1 : self.targetIn, 
-            2 : self.targetOut
-        }
-
-        # Stores data needed for calculating size of the on-screen widgets
-        self.__model = sc.SearchModel()
-        # Stores data relating to the algorithm 
-        self.__dataModel = sc.SearchDataModel() 
-        # Handles the logic 
-        self.__controller = sc.SearchController(self, self.__model, self.__dataModel) 
-        # Adds reference to the controller 
-        self.__dataModel.addController(self.__controller) 
-        # Thread algorithm runs in
-        self.__algorithmThread = None
-
-        # Creating and displaying options
-        self.__createOptions() 
-        # Sets the default delay
-        self.__setDelay()
-
-    # This functions handles creating and displaying the options the user is presented with
-    def __createOptions(self) -> None: 
-        self.__createAlgorithmOptions()               
-        self.__createSpeedAdjuster()
-        self.__createArrayAdjuster()
-        self.__createSortShuffleButtons()
-        self.__createTargetAdjuster()
-        self.__createStopSolveButtons()
-
-    # Creates a combo box which displays all algorithms 
-    def __createAlgorithmOptions(self) -> None:
-        #combo box, allows the user to choose what algorithm they want
-        self.__algorithmOptions = ttk.Combobox(self.getOptionsWidgetFrame(), textvariable = tk.StringVar(), state = "readonly", font = (self.getFont(), 12),\
-             width = self.getOptionsWidgetFrame().winfo_width())
-        self.__algorithmOptions['value'] = getAlgorithms()
-        self.__algorithmOptions.set('Select an algorithm.')
-        # Removes the blue highlighting when something is selected that annoyed me
-        self.__algorithmOptions.bind("<<ComboboxSelected>>", lambda _: self.getOptionsWidgetFrame().focus())
-        self.__algorithmOptions.pack(pady = (10,0)) 
+        self.createBaseLayout()
     
-    # Creates a slider that allows users to adjust an algorithms speed
-    def __createSpeedAdjuster(self) -> None:
-        # Creates a slider that goes 0 to 1 then 2
-        # It has three options correlating to the three speeds; slow, medium, fast 
-        # Every time the sliders value is changed the setDelay() method is called
-        self.__speedSlider = tk.Scale(self.getOptionsWidgetFrame(), from_ = self.__model.getMaxDelay(), to_ = self.__model.getMinDelay(), resolution=0.1, 
-                                      length = self.getOptionsWidgetFrame().winfo_width(), orient = "horizontal", showvalue = False, 
-                                      bg =  "white", highlightbackground = "white", command = self.__updateDelay)
-        self.__speedSlider.pack(pady = (10, 0))  
-        self.__speedSlider.set(self.__model.getMaxDelay())
-        self.__speedSlider.bind("<ButtonRelease-1>", lambda _ : self.__setDelay())
-        # Initially the slider is set at 0, which is the Slow speed
-        #self.__speedSlider.config(label = f"Slow: {self.__speedSlider.get()} seconds")  
-    
-    # Creates a slider that allows users to alter an arrays size
-    def __createArrayAdjuster(self) -> None:
-        self.__arraySizeSlider = tk.Scale(self.getOptionsWidgetFrame(), from_ = 1, to_ = self.__model.getMaxBars(), length = self.getOptionsWidgetFrame().winfo_width(),\
-            orient = "horizontal", bg = "white", highlightbackground = "white", command = self.__controller.adjustArray)
-        self.__arraySizeSlider.pack(pady = (10, 0))
+        # Controller and Model classes
+        self.__searchModel = SearchModel()
+        self.__searchController = SearchController(self, self.__searchModel, self.getDataModel()) 
 
+        # Create Target slider
+        self.__createTargetAdjuster() 
+    
     # Creates a slider that lets sers decide if the target is in the array, not in the array or randomly generated
     def __createTargetAdjuster(self) -> None:
         # Creates a slider that goes from 0 to 1 to 2
         # The three values correlate to the three possible target options
         # The target can guranteed to be in the array, guaranteed to not be in the array or randomly selected
         self.__targetSlider = tk.Scale(self.getOptionsWidgetFrame(), from_ = 0, to_ = 2, length = self.getOptionsWidgetFrame().winfo_width(),\
-            orient = "horizontal", bg = "white", highlightbackground = "white", showvalue = False, command = self.intToText)
+            orient = "horizontal", bg = "white", highlightbackground = "white", showvalue = False, command = self.updateDisplayedText)
         self.__targetSlider.pack(pady = (10, 0))
         # Initially the slider is set at 0, which is the target being randomly selected
         self.__targetSlider.config(label = "Target: Random") 
-    
-    def __createSortShuffleButtons(self): 
-        self.__arraySortShuffleFrame = tk.Frame(self.getOptionsWidgetFrame(), bg = "white") 
-        self.__arraySortShuffleFrame.pack(pady=(20, 0)) 
-
-        self.__sortButton = tk.Button(self.__arraySortShuffleFrame, text="Sort.", width = 7, relief = "solid", font = (self.getFont(), 12), command=self.__sortArray)
-        self.__sortButton.grid(row = 0, column = 0, padx = (0,5)) 
-        self.__shuffleButton = tk.Button(self.__arraySortShuffleFrame, text="Shuffle.", width = 7, relief = "solid", font = (self.getFont(), 12), command=self.__shuffleArray)
-        self.__shuffleButton.grid(row = 0, column = 1, padx = (0,5)) 
-    
-    # Creates buttons that lets user execute algorithms or stop them
-    def __createStopSolveButtons(self) -> None:
-        # Frame to store stop and solve buttons in a grid layout
-        algorithmToggleFrame = tk.Frame(self.getOptionsWidgetFrame(), bg = "white")
-        algorithmToggleFrame.pack(side = "bottom", pady = (0,5))
-        # Allows user to see the algorithm in action
-        self.__solveStopButton = tk.Button(algorithmToggleFrame, text = "Solve.", width = 7, relief = "solid", 
-                                           font = (self.getFont(), 12), command = lambda: self.__initAlgorithm())
-        self.__solveStopButton.grid(row = 0, column = 0, padx = (0,5)) 
-        # Allows user to stop algorithm whilst it's running - button is initially disabled
-        self.__pauseResumeButton = tk.Button(algorithmToggleFrame, text = "Pause.", width = 7, relief = "solid", 
-                                             font = (self.getFont(), 12), state = "disabled", command = lambda : self.__pauseAlgorithm())
-        self.__pauseResumeButton.grid(row = 0, column = 1)  
-
-    # When the slider has changed value a label is added with the relevant speed 
-    # The delay is also changed in the DataModel Object
-    def __updateDelay(self, value : str) -> None: 
-        self.__speedSlider.config(label = f"Delay: {value} Seconds")  
-        #self.__dataModel.setDelay(self.__getDelay())
 
     # When the target slider has changed value a label is added to show the relevant target information 
-    def intToText(self, value : str) -> None:
-        self.__targetSlider.config(label = self.__numbersToText[int(value)]) 
-               
-    # Gets options user has selected from the slider (an intger used as the dictionary key)
-    # and calls the paired function stored in the dictionary 
-    # Each function returns an integer -> the target
-    def generateTarget(self) -> int:
-        return self.__generateTargetFunctions[self.__targetSlider.get()]()
-    
-    # Makes sure that target generated has (almost) equal chance to be in the array or not 
-    def targetRandom(self) -> int: 
-        # Generates decimal between 0 and 1 
-        # If decimal is less than or equal to 0.5 make the target in the array 
-        # Gives a roughly 50-50 chance for target to be in the array or out the array
-        if(random.random() <= 0.5): return self.targetIn()
-        # Else call function to generate the target so it is not in the array
-        else: return self.targetOut()
-    
-    # Guarantees target is in the array
-    def targetIn(self) -> int: 
-       # Randomly chooses index from array and returns integers at that index
-       return self.__dataModel.getArray()[random.randint(0, len(self.__dataModel.getArray()) - 1)] 
+    def updateDisplayedText(self, value : str) -> None: 
+        self.__targetSlider.config(label = self.__searchController.updateSliderText(value)) 
 
-    # Guarantees target is not in array
-    def targetOut(self) -> int: 
-        # Chooses a number between the range of arrays smallest value - 20 and arrays largest value + 20
-        target = random.randint(min(self.__dataModel.getArray()) - 20, max(self.__dataModel.getArray()) + 20)
-        # If generated number in array recall function
-        if target in self.__dataModel.getArray(): self.targetOut()
-        # If generated number not in array then just return value
-        else: return target
-        
-    # Call algorithm user has selected
-    def __initAlgorithm(self) -> None: 
-        # Doesn't do anything if user hasn't chosen an algorithm
-        if(self.__getAlgorithmChoice() == 'Select an algorithm.'): 
-            self.__algorithmOptions.config(foreground = "red")
-        else:
-            self.__solveToStop()
-            # Sets flag indicating the algorithm needs to halt to false
-            if(self.__dataModel.isStopped()): self.__dataModel.clearStopFlag()
-            self.__widgetsAlgorithmStarts()
-            # Generates target the algorithm looks for 
-            self.__dataModel.setTarget(self.generateTarget())
-            # Call algorithm -> so this program actually has a use
-            self.__algorithmThread = threading.Thread(target=callAlgorithm, args=(self.__dataModel, self.__getAlgorithmChoice(), 
-                                                                                  self.__widgetsAlgorithmStops))
-            # Start Thread
-            self.__algorithmThread.start()
-    
-    # Forces current running algorithm thread to terminate (safely)
-    def __stopAlgorithm(self):
-        # Sets to falg to True -> this is what tells the thread/s to stop
-        self.__dataModel.setStopFlag()   
-        # If the algorithm has been paused
-        if(self.__dataModel.isPaused()):
-            # Tell algorithm to resume, so it can stop...
-            self.__resumeAlgorithm()  
-        # Otherwise makes sure pause/resume button has the correct text/function
-        else: 
-            self.__resumeToPause() 
-        self.__widgetsAlgorithmStops()
-    
-    # Returns algorithm the user has selected 
-    def __getAlgorithmChoice(self) -> str:
-        return self.__algorithmOptions.get() 
-    
-    def __setDelay(self): 
-        self.__dataModel.setDelay(self.__speedSlider.get()) 
-  
-    # Changes solve button text and function it calls when it's pressed
-    def __solveToStop(self):
-        self.__solveStopButton.config(text="Stop.", command=self.__stopAlgorithm)
-    
-    # Changes stop button text and function it calls when it's pressed
-    def __stopToSolve(self):
-        self.__solveStopButton.config(text="Solve.", command=self.__initAlgorithm) 
-    
-    # Changes pause button text and function it calls when it's pressed
-    def __pauseToResume(self): 
-         self.__pauseResumeButton.config(text="Resume.", command=self.__resumeAlgorithm) 
-
-    # Changes resume button text and function it calls when it's pressed    
-    def __resumeToPause(self):
-        self.__pauseResumeButton.config(text="Pause.", command=self.__pauseAlgorithm)
-        
-    # Holds the lock, pausing the algorithm Thread
-    def __pauseAlgorithm(self):
-        self.__dataModel.acquireLock() 
-        self.__pauseToResume()
-       
-    # Releases the lock, letting the algorithm thread run again
-    def __resumeAlgorithm(self): 
-        self.__dataModel.releaseLock()
-        self.__resumeToPause()
-    
-    # Enables the button to pause/resume algorithm
-    def __enablePauseResumeButton(self):
-        self.__pauseResumeButton.config(state="active")
-    
-    # Disables the button to pause/resume algorithm
-    def __disablePauseResumeButton(self):
-        self.__pauseResumeButton.config(state="disabled") 
-    
-    # Changes the function of the home buttton
-    def __overrideHomeButtonCommand(self):
-        self.getHomeButton().config(command=self.__loadHomeScreen) 
-    
-    # Ensures any algorithm threads are terminated before moving to the homescreen
-    def __loadHomeScreen(self): 
-        # If a thread exists and it is still running
-        if(self.__algorithmThread and self.__algorithmThread.is_alive()): 
-            # Tell the thread to stop
-            self.__stopAlgorithm() 
-            # Loop until thread has stopped
-            while(self.__algorithmThread.is_alive()): continue 
-        # Load home screen
-        self.loadHomeScreen() 
-    
-    # Sorts and displays the array
-    def __sortArray(self):
-        self.__dataModel.sortArray()
-        self.__controller.displayArray() 
-    
-    # Enables the button to sort the array
-    def __enableSortButton(self):
-        self.__sortButton.config(state="active")
-    
-    # Disables the button to sort the array
-    def __disableSortButton(self):
-        self.__sortButton.config(state="disabled")
-    
-    # Shuffles and displays the array
-    def __shuffleArray(self):
-        self.__dataModel.shuffleArray()
-        self.__controller.displayArray()  
-    
-    # Enables the button to shuffle the array
-    def __enableShuffleButton(self):
-        self.__shuffleButton.config(state="active")
-    
-    # Disables the button to shuffle the array
-    def __disableShuffleButton(self):
-        self.__shuffleButton.config(state="disabled")
-    
-    # Enables the slider to change the array size
-    def __enableArraySizeSlider(self):
-        self.__arraySizeSlider.config(state="active")
-    
-    # Disables the slider to change the array size
-    def __disableArraySizeSlider(self):
-        self.__arraySizeSlider.config(state = "disabled")
-    
-    # Enables/Disables widgets when algorithms runs
-    def __widgetsAlgorithmStarts(self):
-        self.__disableArraySizeSlider()
-        self.__disableSortButton()
-        self.__disableShuffleButton() 
-        self.__solveToStop()
-        self.__enablePauseResumeButton()
-    
-    # Enables/Disables widgets when algorithms stops
-    def __widgetsAlgorithmStops(self):
-        self.__enableArraySizeSlider()
-        self.__enableSortButton()
-        self.__enableShuffleButton()  
-        self.__stopToSolve()
-        self.__disablePauseResumeButton()
-
-# Listen to Whatsername by Green Day    
+# Wretches and Kings by Linkin Park
