+++ screens/shared_layout/__init__.py
@@ -1,0 +1,4 @@
+from .shared_layout import SharedLayout
+from .shared_controller import SharedController
+from .shared_model import SharedModel
+from .data_model import SharedDataModel
